+++ cypress/cypress/component/PanelSidebar/PanelSidebar.cy.tsx
@@ -1,6 +1,13 @@
-import React, { PropsWithChildren, ReactNode } from "react";
-import { PanelSideBarProvider, PanelSideBarLayout, PanelItem, PanelLinkRendererProps, usePanelSideBarContext } from "react-pattern-ui";
-import { faBars, faCogs, faInfo } from "@fortawesome/free-solid-svg-icons";
+import React, { PropsWithChildren } from "react";
+import {
+  PanelSideBarProvider,
+  PanelSideBarLayout,
+  PanelItem,
+  PanelLinkRendererProps,
+  usePanelSideBarContext,
+  PanelItemOnSideBarCollapseOptions,
+} from "react-pattern-ui";
+import { faBars, faCogs, faInfo, faHome, faPerson } from "@fortawesome/free-solid-svg-icons";
 
 type AppRoutes = "home" | "settings" | "dropdownTest" | "dropdown-test1" | "dropdown-test2" | "info";
 type TSideBarMenuItem = PanelItem<AppRoutes>;
@@ -27,7 +34,7 @@ const getPanelSidebarInternal = (items: TSideBarMenuItem[], config?: PanelSideBa
             }
           }}
         >
-          <>{elem.item.title}</>
+          <>{elem.children}</>
         </div>
       )}
     >
@@ -38,17 +45,31 @@ const getPanelSidebarInternal = (items: TSideBarMenuItem[], config?: PanelSideBa
   );
 };
 
-const getSidebarItems = (active?: boolean, disabled?: boolean, expanded?: boolean): TSideBarMenuItem[] => [
+const getSidebarItems = (
+  active?: boolean,
+  disabled?: boolean,
+  expanded?: boolean,
+  onSidebarCollapseOptions?: PanelItemOnSideBarCollapseOptions,
+): TSideBarMenuItem[] => [
   {
     id: "home",
     title: "Home",
     icon: faBars,
     disabled,
+    onSidebarCollapseOptions: onSidebarCollapseOptions ? { ...onSidebarCollapseOptions } : undefined,
     children: [
       {
         title: "Home",
         id: "home",
         active,
+        icon: faHome,
+      },
+      {
+        title: "Profile",
+        id: "profile",
+        onSidebarCollapseOptions: {
+          fallbackIcon: faPerson,
+        },
       },
     ],
   },
@@ -99,11 +120,12 @@ interface PanelSideBarProps extends PropsWithChildren {
   active?: boolean;
   disabled?: boolean;
   expanded?: boolean;
+  onSidebarCollapseOptions?: PanelItemOnSideBarCollapseOptions;
 }
 
 const PanelSideBarWithTiles = (props: PanelSideBarProps) => {
-  const { active, disabled, expanded, children } = props;
-  return getPanelSidebarInternal(getSidebarItems(active, disabled, expanded), { children });
+  const { active, disabled, expanded, onSidebarCollapseOptions, children } = props;
+  return getPanelSidebarInternal(getSidebarItems(active, disabled, expanded, onSidebarCollapseOptions), { children });
 };
 
 const PanelSideBarNoTiles = (props: PanelSideBarProps) => {
@@ -266,5 +288,16 @@ describe("PanelSidebar.cy.tsx", () => {
     cy.get("button[title=Settings]").should("be.visible");
     cy.get("button[title=Home]").should("be.visible");
     cy.get("button[title=Info]").should("not.exist");
+  });
+
+  it("toggle sidebar with visible icons", () => {
+    cy.mount(<PanelSideBarWithTiles onSidebarCollapseOptions={{ showIcon: true }} />);
+    cy.get('[data-icon="angle-left"]').should("be.visible");
+    cy.get("#side-nav-toggle").click();
+    cy.get('[data-icon="angle-right"]').should("be.visible");
+    cy.get(".toggled").should("exist");
+    cy.get(".side-nav__items").should("be.visible");
+    cy.get("#home").should("be.visible");
+    cy.get("#profile > .nav-link > svg").should("be.visible");
   });
 });