+++ src/lib/Layout/PanelSideBarLayout/PanelSideBar/Context/PanelSideBarContext.tsx
@@ -87,6 +87,8 @@ export const PanelSideBarProvider = <TPanelItemId extends string, TPanelItem>(
     );
   };
 
+  const isIconShownOnSidebarCollapse = menuItems.find((x) => x.id === activePanelId)?.onSidebarCollapseOptions?.showIcon || false;
+
   return (
     <PanelSideBarContext.Provider
       value={{
@@ -106,6 +108,7 @@ export const PanelSideBarProvider = <TPanelItemId extends string, TPanelItem>(
         closeMenuItems,
         hiddenMenuItemIds,
         setHiddenMenuItemsIds,
+        isIconShownOnSidebarCollapse,
       }}
     >
       {children}